put classifier into single file
